Update sync API version from v8 to v9
I am aware that this package is now deprecated and only the RESTful python client is now supported. But updating the endpoint from v8 to v9 allowed me to continue to use this package as a dependency to todoist-habitica-sync. Would appreciate if this can be kept alive with minimal changes.